requirements.txt
appdirs==1.4.4
bcrypt==3.1.7
cffi==1.15.1
cryptography==2.8
exceptiongroup==1.2.2
Fabric3==1.14.post1
Flask==2.2.5
Flask-Cors==4.0.1
Flask-HTTPAuth==4.8.0
greenlet==3.0.3
importlib-metadata==6.7.0
iniconfig==2.0.0
itsdangerous==2.1.2
mysqlclient==2.1.1
packaging==24.0
paramiko==2.12.0
pluggy==1.2.0
pycparser==2.21
PyNaCl==1.3.0
pyparsing==3.1.2
pytest==7.4.4
six==1.16.0
SQLAlchemy==2.0.31
tomli==2.0.1
typing_extensions==4.7.1
Werkzeug==2.2.3
zipp==3.15.0
